test(core): align user tests with owners
