+++ cli/internal/core/users_test.go
@@ -3,7 +3,6 @@ package core
 import (
 	"context"
 	"errors"
-	"sync"
 	"testing"
 	"time"
 
@@ -268,168 +267,6 @@ func TestChattoCore_ConcurrentUserCreation(t *testing.T) {
 	}
 	if user2 != "" && retrieved.Id != user2 {
 		t.Errorf("Expected user ID %s, got %s", user2, retrieved.Id)
-	}
-}
-
-func TestChattoCore_LinkExternalIdentity(t *testing.T) {
-	core, _ := setupTestCore(t)
-
-	ctx := context.Background()
-	user, err := core.CreateUser(ctx, "system", "externaluser", "External User", "password123")
-	if err != nil {
-		t.Fatalf("CreateUser: %v", err)
-	}
-	other, err := core.CreateUser(ctx, "system", "externalother", "External Other", "password123")
-	if err != nil {
-		t.Fatalf("CreateUser other: %v", err)
-	}
-
-	if err := core.LinkExternalIdentity(ctx, "github-main", "github", "github-main", "12345", user.Id); err != nil {
-		t.Fatalf("LinkExternalIdentity: %v", err)
-	}
-	if err := core.LinkExternalIdentity(ctx, "github-main", "github", "github-main", "12345", user.Id); err != nil {
-		t.Fatalf("LinkExternalIdentity idempotent: %v", err)
-	}
-
-	found, err := core.GetUserByExternalIdentity(ctx, "github-main", "12345")
-	if err != nil {
-		t.Fatalf("GetUserByExternalIdentity: %v", err)
-	}
-	if found == nil || found.Id != user.Id {
-		t.Fatalf("GetUserByExternalIdentity = %v, want %s", found, user.Id)
-	}
-
-	err = core.LinkExternalIdentity(ctx, "github-main", "github", "github-main", "12345", other.Id)
-	if !errors.Is(err, ErrExternalIdentityAlreadyClaimed) {
-		t.Fatalf("LinkExternalIdentity conflict error = %v, want ErrExternalIdentityAlreadyClaimed", err)
-	}
-}
-
-func TestChattoCore_DisconnectExternalIdentity(t *testing.T) {
-	core, _ := setupTestCore(t)
-
-	ctx := context.Background()
-	user, err := core.CreateUser(ctx, "system", "disconnectuser", "Disconnect User", "password123")
-	if err != nil {
-		t.Fatalf("CreateUser: %v", err)
-	}
-	if err := core.LinkExternalIdentity(ctx, "github-main", "github", "github-main", "12345", user.Id); err != nil {
-		t.Fatalf("LinkExternalIdentity: %v", err)
-	}
-	authGeneration, err := core.CurrentAuthGeneration(ctx, user.Id)
-	if err != nil {
-		t.Fatalf("CurrentAuthGeneration: %v", err)
-	}
-	token, err := core.CreateAuthTokenWithSourceGeneration(ctx, user.Id, "external_identity_login", authGeneration)
-	if err != nil {
-		t.Fatalf("CreateAuthTokenWithSourceGeneration: %v", err)
-	}
-	sessionID, _, err := core.CreateCookieSessionForGeneration(ctx, user.Id, "external_identity_login", authGeneration)
-	if err != nil {
-		t.Fatalf("CreateCookieSessionForGeneration: %v", err)
-	}
-	identities, err := core.ExternalIdentitiesForUser(ctx, user.Id)
-	if err != nil {
-		t.Fatalf("ExternalIdentitiesForUser: %v", err)
-	}
-	if len(identities) != 1 || identities[0].SubjectHash == "" {
-		t.Fatalf("identities = %+v", identities)
-	}
-
-	if err := core.DisconnectExternalIdentity(ctx, user.Id, identities[0].SubjectHash); err != nil {
-		t.Fatalf("DisconnectExternalIdentity: %v", err)
-	}
-	found, err := core.GetUserByExternalIdentity(ctx, "github-main", "12345")
-	if err != nil {
-		t.Fatalf("GetUserByExternalIdentity: %v", err)
-	}
-	if found != nil {
-		t.Fatalf("GetUserByExternalIdentity after disconnect = %v, want nil", found)
-	}
-	identities, err = core.ExternalIdentitiesForUser(ctx, user.Id)
-	if err != nil {
-		t.Fatalf("ExternalIdentitiesForUser after disconnect: %v", err)
-	}
-	if len(identities) != 0 {
-		t.Fatalf("identities after disconnect = %+v, want empty", identities)
-	}
-	if _, err := core.ValidateAuthToken(ctx, token); !errors.Is(err, ErrAuthTokenNotFound) {
-		t.Fatalf("ValidateAuthToken after disconnect err = %v, want ErrAuthTokenNotFound", err)
-	}
-	if _, err := core.ValidateCookieSession(ctx, user.Id, sessionID); !errors.Is(err, ErrCookieSessionNotFound) {
-		t.Fatalf("ValidateCookieSession after disconnect err = %v, want ErrCookieSessionNotFound", err)
-	}
-	if _, err := core.CreateAuthTokenWithSourceGeneration(ctx, user.Id, "external_identity_login", authGeneration); !errors.Is(err, ErrAuthTokenNotFound) {
-		t.Fatalf("CreateAuthTokenWithSourceGeneration old generation err = %v, want ErrAuthTokenNotFound", err)
-	}
-	afterGeneration, err := core.CurrentAuthGeneration(ctx, user.Id)
-	if err != nil {
-		t.Fatalf("CurrentAuthGeneration after disconnect: %v", err)
-	}
-	if afterGeneration == authGeneration {
-		t.Fatal("auth generation should advance after external identity disconnect")
-	}
-
-	err = core.DisconnectExternalIdentity(ctx, user.Id, "missing-subject-hash")
-	if !errors.Is(err, ErrExternalIdentityNotFound) {
-		t.Fatalf("DisconnectExternalIdentity missing error = %v, want ErrExternalIdentityNotFound", err)
-	}
-}
-
-func TestChattoCore_DisconnectExternalIdentityConcurrentDuplicate(t *testing.T) {
-	core, _ := setupTestCore(t)
-
-	ctx := context.Background()
-	user, err := core.CreateUser(ctx, "system", "disconnectrace", "Disconnect Race", "password123")
-	if err != nil {
-		t.Fatalf("CreateUser: %v", err)
-	}
-	if err := core.LinkExternalIdentity(ctx, "github-main", "github", "github-main", "12345", user.Id); err != nil {
-		t.Fatalf("LinkExternalIdentity: %v", err)
-	}
-	identities, err := core.ExternalIdentitiesForUser(ctx, user.Id)
-	if err != nil {
-		t.Fatalf("ExternalIdentitiesForUser: %v", err)
-	}
-	if len(identities) != 1 || identities[0].SubjectHash == "" {
-		t.Fatalf("identities = %+v", identities)
-	}
-
-	start := make(chan struct{})
-	errs := make(chan error, 2)
-	var wg sync.WaitGroup
-	for i := 0; i < 2; i++ {
-		wg.Add(1)
-		go func() {
-			defer wg.Done()
-			<-start
-			errs <- core.DisconnectExternalIdentity(ctx, user.Id, identities[0].SubjectHash)
-		}()
-	}
-	close(start)
-	wg.Wait()
-	close(errs)
-
-	var successes, notFound int
-	for err := range errs {
-		switch {
-		case err == nil:
-			successes++
-		case errors.Is(err, ErrExternalIdentityNotFound):
-			notFound++
-		default:
-			t.Fatalf("DisconnectExternalIdentity concurrent error = %v", err)
-		}
-	}
-	if successes != 1 || notFound != 1 {
-		t.Fatalf("concurrent disconnect results: successes=%d notFound=%d, want 1 each", successes, notFound)
-	}
-	identities, err = core.ExternalIdentitiesForUser(ctx, user.Id)
-	if err != nil {
-		t.Fatalf("ExternalIdentitiesForUser after concurrent disconnect: %v", err)
-	}
-	if len(identities) != 0 {
-		t.Fatalf("identities after concurrent disconnect = %+v, want empty", identities)
 	}
 }
